Category

+++ src/module/categories/categories.controller.ts
@@ -48,16 +48,16 @@ export class CategoriesController {
   @ApiBody({
     schema: {
       type: 'object',
-      required: ['file', 'cat_title', 'cat_description', 'cat_image'],
+      required: ['file', 'title', 'description', 'image'],
       properties: {
         file: {
           type: 'string',
           format: 'binary',
         },
-        cat_title: {
+        title: {
           type: 'string',
         },
-        cat_description: {
+        description: {
           type: 'string',
         },
       },
@@ -67,8 +67,8 @@ export class CategoriesController {
   @ApiBadRequestResponse()
   @UseInterceptors(FileInterceptor('categories'))
   @ApiHeader({
-    name: 'admin_token',
-    description: 'Admin token',
+    name: 'authorization',
+    description: 'Authorization',
     required: true,
   })
   async uploadfile(
@@ -76,10 +76,10 @@ export class CategoriesController {
     @Body() createCategoryDto: CreateCategoryDto,
     @Headers() headers: any,
   ) {
-    const admin = await this.verifyToken.verifyAdmin(headers)
+    const admin = await this.verifyToken.verifyAdmin(headers);
     if (admin) {
       const cat_link: any = googleCloud(file);
-      return this.categoriesService.create(createCategoryDto, cat_link);
+      return await this.categoriesService.create(createCategoryDto, cat_link);
     }
   }
 
@@ -100,12 +100,15 @@ export class CategoriesController {
         file: {
           type: 'string',
           format: 'binary',
+          default: 'dsfsgf'
         },
-        cat_title: {
+        title: {
           type: 'string',
+          default: 'Node.js'
         },
-        cat_description: {
+        description: {
           type: 'string',
+          default: 'Zo`r dasturlash tili'
         },
       },
     },
@@ -114,8 +117,8 @@ export class CategoriesController {
   @ApiBadRequestResponse()
   @UseInterceptors(FileInterceptor('categories'))
   @ApiHeader({
-    name: 'admin_token',
-    description: 'Admin token',
+    name: 'authorization',
+    description: 'Authorization',
     required: true,
   })
   @HttpCode(HttpStatus.NO_CONTENT)
@@ -126,10 +129,10 @@ export class CategoriesController {
     @Body() body: UpdateCategoryDto,
     @Headers() headers: any,
   ) {
-    const admin = await this.verifyToken.verifyAdmin(headers)
+    const admin = await this.verifyToken.verifyAdmin(headers);
     const cat_link: any = googleCloud(file);
     if (admin) {
-      return this.categoriesService.update(id, body, cat_link);
+      return await this.categoriesService.update(id, body, cat_link);
     }
   }
 
@@ -138,14 +141,16 @@ export class CategoriesController {
   @ApiNotFoundResponse()
   @ApiUnprocessableEntityResponse()
   @ApiForbiddenResponse()
-  @Delete('/delete/:id') 
+  @Delete('/delete/:id')
   @ApiHeader({
-    name: 'admin_token',
-    description: 'Admin token',
+    name: 'authorization',
+    description: 'Authorization',
     required: true,
   })
   async remove(@Param('id') id: string, @Headers() headers: any) {
-    const admin = await this.verifyToken.verifyAdmin(headers)
-    return this.categoriesService.remove(id);
+    const admin = await this.verifyToken.verifyAdmin(headers);
+    if (admin) {
+      return await this.categoriesService.remove(id);
+    }
   }
 }
